Base game build
A 3D pong game made from a tutorial from which I will optimize, modify and expand upon. It contains two rackets, able to be controlled by player input or simple AI, a ball which is kinematic and controlled via scripting, and a simple scoring script with a ball reset and animated countdown to start ball movement, as well as a pre-made gameplay field with emissive materials to add visual variety.