Polish update
Added scripts for managing UI and options menu. Organized scripts into folders based on function. Created light and occlusion culling data. Added audiomixer and assigned exposed variables to it, linked to Audio script and buttons/sliders. Created a panel where a player can choose which racket to control. Options are to control 1 manually, control both manually, or have both be controlled by computer. Added music to title screen. Created a variation of scripts for the title screen which allows for the game to play automatically. Added a version display to title screen.